+++ arm_chip_ino.h
@@ -72,11 +72,6 @@ __INLINE FctERR HALERRtoFCTERR(const SDWORD status)
 }
 
 
-#ifndef UNUSED
-#define UNUSED(X)			(void) (X)			//!< To avoid gcc/g++ warnings
-#endif
-
-
 /****************************************************************/
 #ifdef __cplusplus
 	}