+++ arm_chip_pic.h
@@ -85,11 +85,6 @@ __INLINE FctERR HALERRtoFCTERR(const SDWORD status)
 }
 
 
-#ifndef UNUSED
-#define UNUSED(X)			(void) (X)			//!< To avoid gcc/g++ warnings
-#endif
-
-
 /****************************************************************/
 #ifdef __cplusplus
 	}